Use OfflineAudioContext instead of AudioContext to work around autoplay restrictions.
This is a proof of concept and could most likely benefit from memory and
performance improvements. On Chrome Dev, after a couple of rotations,
there is a visible slow down when switching OfflineAudioContext.

This proof of concept may make a case for an OfflineAudioContext that
streams a buffer instead of creating one with a given size.